feat(0000): Add coverage test detail & docs (#5)
* feat(000): Cypress testing for React Application

Patching the package.json file with ENVIRONMENT VAR to run the CI upload of video. Later this will be provided by the github actions.

+ fix: command cypress:run add ENVIRONMENT VAR 'CYPRESS_RECORD_KEY' to upload video

Signed-off-by: Kyle Jeske <kylejeske@users.noreply.github.com>

* feat(000): Cypress testing for React Application

Patching the package.json file with ENVIRONMENT VAR to run the CI upload of video. Later this will be provided by the github actions.

+ fix: command cypress:run add ENVIRONMENT VAR 'CYPRESS_RECORD_KEY' to upload video
+ add: Details of the cypress actions files
+ del: Coverage files, folders name from .gitignore
+ add: Include coverage data files from tests

Signed-off-by: Kyle Jeske <kylejeske@users.noreply.github.com>

* feat(0000): Include coverage tests files

Since the file aren't included withing the gitignore, include coverage report files.

+ add: Include coverage folder with lcov-report subfolder

Signed-off-by: Kyle Jeske <kylejeske@users.noreply.github.com>

* feat(0000): Add coverage tests into example